+++ packages/components/src/navigation-drawer/navigation-drawer.ts
@@ -1,16 +1,16 @@
 /* eslint-disable max-len */
-import { GecutDirective } from '@gecut/lit-helper/directives/directive.js';
-import { mapObject } from '@gecut/lit-helper/utilities/map-object.js';
-import { map } from '@gecut/lit-helper/utilities/map.js';
-import { directive, type PartInfo } from 'lit/directive.js';
-import { ifDefined } from 'lit/directives/if-defined.js';
-import { when } from 'lit/directives/when.js';
-import { html, noChange } from 'lit/html.js';
+import {GecutDirective} from '@gecut/lit-helper/directives/directive.js';
+import {mapObject} from '@gecut/lit-helper/utilities/map-object.js';
+import {map} from '@gecut/lit-helper/utilities/map.js';
+import {directive, type PartInfo} from 'lit/directive.js';
+import {ifDefined} from 'lit/directives/if-defined.js';
+import {when} from 'lit/directives/when.js';
+import {html, noChange} from 'lit/html.js';
 
-import { divider } from '../divider/divider.js';
-import { icon } from '../icon/icon.js';
+import {divider} from '../divider/divider.js';
+import {icon} from '../icon/icon.js';
 
-import type { NavigationItemContent } from '../navigation-bar/navigation-bar.js';
+import type {NavigationItemContent} from '../navigation-bar/navigation-bar.js';
 
 export interface NavigationDrawerGroupContent {
   title?: string;
@@ -41,9 +41,7 @@ export class GecutNavigationDrawerDirective extends GecutDirective {
         id="navigationDrawer"
         class="absolute bottom-0 start-0 top-0 z-modal w-[22.5rem] max-w-[88vw] translate-x-[-105%] rtl:translate-x-[105%] transform-gpu overflow-clip rounded-e-3xl md:rounded-3xl md:my-3 bg-surfaceContainerHigh transition-transform duration-300 ease-in will-change-transform elevation-1 [&.opened]:translate-x-0 [&.opened]:ease-out [&.opened]:shadow-[0_0_100vw_100vw_rgba(0,0,0,0.3)] md:[&.opened]:-translate-x-2 md:[&.opened]:elevation-3 border-0 border-surfaceContainerHighest dark:md:border-2 md:my-5 ${openAttr}"
       >
-        <nav
-          class="flex h-full flex-col bg-surfaceContainerLow px-3 py-3 elevation-1"
-        >
+        <nav class="flex h-full flex-col bg-surfaceContainerLow px-3 py-3 elevation-1">
           ${this._renderGroups(content)}
         </nav>
       </aside>
@@ -58,23 +56,16 @@ export class GecutNavigationDrawerDirective extends GecutDirective {
         <div role="tablist" class="text-labelLarge text-onSurfaceVariant">
           ${when(
             groupContent.title,
-            () => html`
-              <h2 class="mx-4 pb-3 pt-5 text-titleSmall text-onSurfaceVariant">
-                ${groupContent.title}
-              </h2>
-            `
+            () => html` <h2 class="mx-4 pb-3 pt-5 text-titleSmall text-onSurfaceVariant">${groupContent.title}</h2> `,
           )}
           ${this._renderItems(groupContent, content.selected)}
-          ${when(groupContent.divider, () => divider({ gapTop: true }))}
+          ${when(groupContent.divider, () => divider({gapTop: true}))}
         </div>
-      `
+      `,
     );
   }
 
-  protected _renderItems(
-    content: NavigationDrawerGroupContent,
-    _selected: string
-  ): unknown {
+  protected _renderItems(content: NavigationDrawerGroupContent, _selected: string): unknown {
     return mapObject(this, content.paths, (item, name) => {
       const isLink = 'path' in item;
       const href = isLink ? item.path : undefined;
@@ -97,12 +88,8 @@ export class GecutNavigationDrawerDirective extends GecutDirective {
           tabindex="0"
           class="flex h-14 cursor-pointer select-none flex-nowrap items-center rounded-full px-3 hover:bg-secondaryContainer hover:text-onSecondaryContainer hover:stateHover-onSecondaryContainer focus-ring [&>.gecut-icon]:mx-1 [&>.gecut-icon]:h-6 [&>.gecut-icon]:w-6 aria-selected:stateActive-onSecondaryContainer aria-selected:text-onSecondaryContainer"
         >
-          ${icon(item.icon)}
-          ${when(label, () => html`<div class="mx-2 grow">${label}</div>`)}
-          ${when(
-            item.badge != null,
-            () => html`<div class="me-3">${item.badge}</div>`
-          )}
+          ${icon(item.icon)} ${when(label, () => html`<div class="mx-2 grow">${label}</div>`)}
+          ${when(item.badge != null, () => html`<div class="me-3">${item.badge}</div>`)}
         </a>
       `;
     });