+++ packages/lit-helper/src/pages/page.ts
@@ -1,8 +1,8 @@
-import { scheduleSignalElement } from '@gecut/mixins';
-import { html } from 'lit';
+import {scheduleSignalElement} from '@gecut/mixins';
+import {html} from 'lit';
 
-import type { RenderResult } from '@gecut/types';
-import type { PropertyValueMap } from 'lit';
+import type {RenderResult} from '@gecut/types';
+import type {PropertyValueMap} from 'lit';
 
 export interface GecutPWAPageMeta {
   title?: string;
@@ -22,9 +22,7 @@ export abstract class GecutPWAPage extends scheduleSignalElement {
     return html``;
   }
 
-  protected override firstUpdated(
-    _changedProperties: PropertyValueMap<this> | Map<PropertyKey, unknown>
-  ): void {
+  protected override firstUpdated(_changedProperties: PropertyValueMap<this> | Map<PropertyKey, unknown>): void {
     super.firstUpdated(_changedProperties);
 
     this.classList.add('gecut-page');