+++ packages/lit-helper/src/utilities/map-filter.ts
@@ -1,4 +1,4 @@
-import type { MaybePromise, RenderResult } from '@gecut/types';
+import type {MaybePromise, RenderResult} from '@gecut/types';
 
 // eslint-disable-next-line max-params, consistent-return
 export async function* mapFilter<T>(
@@ -6,14 +6,14 @@ export async function* mapFilter<T>(
   items: T[] | undefined | null,
   filter: (item: T) => MaybePromise<boolean>,
   f: (value: T) => RenderResult,
-  loading?: RenderResult
+  loading?: RenderResult,
 ): AsyncGenerator<RenderResult, RenderResult, RenderResult> {
   if (items == null) {
     return loading;
   }
 
   for (const value of items) {
-    if (await Promise.resolve(filter(value)) === true) {
+    if ((await Promise.resolve(filter(value))) === true) {
       yield f.call(_this, value);
     }
   }