+++ packages/lit-helper/src/utilities/map.ts
@@ -1,11 +1,11 @@
-import type { RenderResult } from '@gecut/types';
+import type {RenderResult} from '@gecut/types';
 
 // eslint-disable-next-line consistent-return
 export function* map<T>(
   _this: unknown,
   items: T[] | undefined | null,
   f: (value: T) => RenderResult,
-  loading?: RenderResult
+  loading?: RenderResult,
 ): Generator<RenderResult, RenderResult, RenderResult> {
   if (items == null) {
     return loading;