+++ packages/mixins/src/mixins.ts
@@ -1,9 +1,9 @@
 /* eslint-disable new-cap */
-import { LitElement } from 'lit';
+import {LitElement} from 'lit';
 
-import { LoggerMixin } from './lib/logger.js';
-import { ScheduleUpdateToFrameMixin } from './lib/schedule-update-to-frame.js';
-import { SignalMixin } from './lib/signal.js';
+import {LoggerMixin} from './lib/logger.js';
+import {ScheduleUpdateToFrameMixin} from './lib/schedule-update-to-frame.js';
+import {SignalMixin} from './lib/signal.js';
 
 export * from './lib/logger.js';
 export * from './lib/signal.js';