+++ packages/lit-helper/CHANGELOG.md
@@ -11,8 +11,8 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Bug Fixes
 
-* **lit-helper:** eslint issues ([82addf0](https://github.com/gecut/hybrid-ui/commit/82addf08e6f839659082b6a9313b0cbc1680fc66)) by @MM25Zamanian
+- **lit-helper:** eslint issues ([82addf0](https://github.com/gecut/hybrid-ui/commit/82addf08e6f839659082b6a9313b0cbc1680fc66)) by @MM25Zamanian
 
 ### Features
 
-* **lit-helper:** new packages ([f2ac41c](https://github.com/gecut/hybrid-ui/commit/f2ac41c9dcb2f2ac0ba48b01b6661c4f9ab70dcd)) by @MM25Zamanian
+- **lit-helper:** new packages ([f2ac41c](https://github.com/gecut/hybrid-ui/commit/f2ac41c9dcb2f2ac0ba48b01b6661c4f9ab70dcd)) by @MM25Zamanian